API

Hilbert Spaces

q-Numbers

Symbolic Numbers

Average

Utility functions

Symbolic Summations

Hilbert Spaces

SecondQuantizedAlgebra.HilbertSpace — Type

HilbertSpace

Abstract type for representing Hilbert spaces.

SecondQuantizedAlgebra.ProductSpace — Type

ProductSpace <: HilbertSpace

Stores a composite HilbertSpace consisting of multiple subspaces. Generally created by computing the tensor product ⊗ of subspaces.

SecondQuantizedAlgebra.FockSpace — Type

FockSpace <: HilbertSpace

HilbertSpace defining a Fock space for bosonic operators. See also: Destroy, Create

SecondQuantizedAlgebra.NLevelSpace — Type

NLevelSpace <: HilbertSpace
NLevelSpace(name::Symbol,levels,GS=levels[1])

Define a HilbertSpace for an object consisting of N discrete energy levels. The given levels must be an integer specifying the number of levels, or an iterable collection of levels. The argument GS specifies which state should be treated as ground state and is rewritten using population conservation during simplification. See also: Transition

Examples:

julia> ha = NLevelSpace(:a,3)
ℋ(a)

julia> ha = NLevelSpace(:a,(:g,:e))
ℋ(a)

SecondQuantizedAlgebra.ClusterSpace — Type

ClusterSpace <: HilbertSpace
ClusterSpace(original_space,N,order)

A Hilbert space representing N identical copies of another Hilbert space, with correlations up to a specified order.

QuantumInterface.:⊗ — Function

⊗(spaces::HilbertSpace...)

Create a ProductSpace consisting of multiple subspaces. Unicode \otimes<tab> alias of tensor

Examples

julia> hf = FockSpace(:f)
ℋ(f)

julia> ha = NLevelSpace(:a,2)
ℋ(a)

julia> h = hf⊗ha
ℋ(f) ⊗ ℋ(a)

QuantumInterface.tensor — Function

tensor(spaces::HilbertSpace...)

Create a ProductSpace consisting of multiple subspaces. See also ⊗.

SecondQuantizedAlgebra.PauliSpace — Type

PauliSpace <: HilbertSpace

HilbertSpace defining a Spin space for two-level atom Pauli operators. See also: Pauli, Create

SecondQuantizedAlgebra.Pauli — Type

Pauli <: QSym

Pauli operator on a PauliSpace representing the Pauli operators σx, σy and σz for two-level spin systems. The field axis represents x, y and z as 1, 2 and 3, respectively. The used rewriting rule is σj⋅σk → δjk + i⋅ϵjkl⋅σl.

Examples

julia> h = PauliSpace("Spin-1/2")
ℋ(Spin-1/2)

julia> σx = Pauli(h,:σ,1)
σx

SecondQuantizedAlgebra.SpinSpace — Type

SpinSpace <: HilbertSpace

HilbertSpace defining a Spin space for N > 1 identical two-level atom operators. See also: Spin, Create

SecondQuantizedAlgebra.Spin — Type

Spin <: QSym

Spin operator on a SpinSpace representing the Spin-operators Sx, Sy and Sz for collective spin systems. The field axis represents x, y and z as 1, 2 and 3, respectively. The operators follow the rules for angular momentum operators: [Sj,Sk] = i⋅∑ϵjkl⋅Sl

Examples

julia> h = SpinSpace("Spin-N/2")
ℋ(Spin-N/2)

julia> Sx = Spin(h,:S,1)
Sx

q-Numbers

SecondQuantizedAlgebra.QSym — Type

QSym <: QNumber

Abstract type representing fundamental operator types.

SecondQuantizedAlgebra.QTerm — Type

QTerm <: QNumber

Abstract type representing noncommutative expressions.

SecondQuantizedAlgebra.@qnumbers — Macro

@qnumbers

Convenience macro for the construction of operators.

Examples

julia> h = FockSpace(:fock)
ℋ(fock)

julia> @qnumbers a::Destroy(h)
(a,)

julia> h = FockSpace(:one) ⊗ FockSpace(:two)
ℋ(one) ⊗ ℋ(two)

julia> @qnumbers b::Destroy(h,2)
(b,)

SecondQuantizedAlgebra.Destroy — Type

Destroy <: QSym

Bosonic operator on a FockSpace representing the quantum harmonic oscillator annihilation operator.

SecondQuantizedAlgebra.Create — Type

Create <: QSym

Bosonic operator on a FockSpace representing the quantum harmonic oscillator creation operator.

SecondQuantizedAlgebra.Transition — Type

Transition <: QSym
Transition(h::NLevelSpace,name::Symbol,i,j)

Fundamental operator defining a transition from level j to level i on a NLevelSpace. The notation corresponds to Dirac notation, i.e. the above is equivalent to |i⟩⟨j|.

Examples

julia> ha = NLevelSpace(:a,(:g,:e))
ℋ(a)

julia> σ = Transition(ha,:σ,:g,:e)
σge

Symbolic Numbers

SecondQuantizedAlgebra.CNumber — Type

CNumber <: Number

Abstract type for all symbolic numbers, i.e. Parameter, average.

SecondQuantizedAlgebra.Parameter — Type

Parameter <: CNumber

Type used as symbolic type in a SymbolicUtils.Sym variable to represent a parameter.

SecondQuantizedAlgebra.cnumbers — Function

cnumbers(symbols::Symbol...)
cnumbers(s::String)

Create symbolic cnumbers.

Expamples

julia> ps = cnumbers(:a, :b)
(a, b)

julia> cnumbers("a b") == ps
true

SecondQuantizedAlgebra.cnumber — Function

cnumber(symbols::Symbol)
cnumber(s::String)

Create symbolic cnumber.

Expamples

julia> ps = cnumber(:a)
a

julia> cnumber("a") == ps
true

SecondQuantizedAlgebra.@cnumbers — Macro

@cnumbers(ps...)

Convenience macro to quickly define symbolic cnumbers.

Examples

julia> @cnumbers ω κ
(ω, κ)

SecondQuantizedAlgebra.RNumber — Type

RNumber <: Real

Abstract type for real symbolic numbers RealParameter.

SecondQuantizedAlgebra.RealParameter — Type

RealParameter <: RNumber

Type used as symbolic type in a SymbolicUtils.Sym variable to represent a real parameter.

SecondQuantizedAlgebra.rnumbers — Function

rnumbers(symbols::Symbol...)
rnumbers(s::String)

Create symbolic rnumbers.

Expamples

julia> ps = rnumbers(:a, :b)
(a, b)

julia> rnumbers("a b") == ps
true

SecondQuantizedAlgebra.rnumber — Function

rnumber(symbols::Symbol)
rnumber(s::String)

Create symbolic rnumber.

Expamples

julia> ps = rnumber(:a)
a

julia> rnumber("a") == ps
true

SecondQuantizedAlgebra.@rnumbers — Macro

@rnumbers(ps...)

Convenience macro to quickly define symbolic rnumbers.

Examples

julia> @rnumbers ω κ
(ω, κ)

Average

SecondQuantizedAlgebra.average — Function

average(::QNumber)

Compute the average of an operator.

Utility functions

SecondQuantizedAlgebra.commutator — Function

commutator(a,b)

Computes the commutator a*b - b*a.

SecondQuantizedAlgebra.acts_on — Function

acts_on(op)

Shows on which Hilbert space op acts. For QSym types, this returns an Integer, whereas for a Term it returns a Vector{Int} whose entries specify all subspaces on which the expression acts.

SecondQuantizedAlgebra.find_operators — Function

find_operators(::HilbertSpace, order; names=nothing)

Find all operators that fully define a system up to the given order.

SecondQuantizedAlgebra.unique_ops — Function

unique_ops(ops)

For a given list of operators, return only unique ones taking into account their adjoints.

SecondQuantizedAlgebra.unique_ops! — Function

unique_ops!(ops)

In-place version of unique_ops.

SecondQuantizedAlgebra.fundamental_operators — Function

fundamental_operators(::HilbertSpace)

Return all fundamental operators for a given Hilbertspace. For example, a FockSpace only has one fundamental operator, Destroy.

SecondQuantizedAlgebra.transition_superscript — Function

transition_superscript(::Bool)

Specify whether the indices in a Transition operator should be printed as superscript. Default is true. If set to false, the indices corresponding to the levels are printed as subscript.

SecondQuantizedAlgebra.to_numeric — Function

to_numeric(q::QNumber, b::QuantumOpticsBase.Basis; level_map = nothing)
to_numeric(q::QNumber, state; level_map = nothing)

Convert a symbolic operator q to its equivalent numeric (matrix) form on the basis b. The optional argument level_map can be set to a dictionary that specifies how to map levels of a Transition to the ones given in an NLevelBasis. Note: If the levels of a transition are symbolic, setting level_map is required.

See also: numeric_average

Examples

julia> to_numeric(Destroy(FockSpace(:fock), :a), FockBasis(10)) Operator(dim=11x11) basis: Fock(cutoff=10)[...]

to_numeric(q::QNumber, b::QuantumOpticsBase.Basis, d::Dict)
to_numeric(q::QNumber, state, d::Dict)

Map a symbolic operator q to a numeric (matrix) defined in the dictionary d. See also: numeric_average

SecondQuantizedAlgebra.numeric_average — Function

numeric_average(avg::Average, state; level_map = nothing)
numeric_average(q::QNumber, state; level_map = nothing)

From a symbolic average avg or operator q, compute the corresponding numerical average value with the given quantum state state. This state can either be of type QuantumOpticsBase.StateVector or QuantumOpticsBase.Operator.

See also: to_numeric

numeric_average(avg::Average, state, d::Dict)
numeric_average(q::QNumber, state, d::Dict)

From a symbolic average avg or operator q, map the operator defined in the dictionary d and compute the numerical average value for an with the given quantum state state. This state can either be of type QuantumOpticsBase.StateVector or QuantumOpticsBase.Operator.

See also: to_numeric

Symbolic Summations

SecondQuantizedAlgebra.Index — Type

Index(hilb::HilbertSpace,name::Symbol,range::Union{Int64,Sym},aon::Int)

Defines an index, using a Symbol as a name, and a HilbertSpace for computation and commutator-relations. Indices with all same fields will be considered equal. See also: IndexedOperator and IndexedVariable

Fields:

hilb: The whole HilbertSpace, the index will be defined on.

name: A Symbol, which defines the name of the index, and how product-terms of IndexedOperator are ordered (alphabetical)

range: The upper bound limit of the index. This can be a SymbolicUitls.Symbolic or any Number.

aon: Number specifying the specific HilbertSpace, where the Index acts on.

SecondQuantizedAlgebra.IndexedOperator — Type

IndexedOperator <: QSym
IndexedOperator(op::Union{Transition,Create,Destroy},ind::Index)

Operator, associated with an index.

Fields:

op: Operator, either a Transition, a Destroy or a Create can be defined.

ind: The index, the operator will be associated with.

SecondQuantizedAlgebra.IndexedVariable — Type

IndexedVariable <: CNumber
IndexedVariable(name::Symbol,ind::Index)
IndexedVariable(name::Symbol,ind1::Index,ind2:Index)

A indexed symbolic variable. The variable can (once equations are calculated) be easily exchanged for numerical values. Calling a IndexedVariable using two different Index objects one can create DoubleIndexedVariable objects.

SecondQuantizedAlgebra.DoubleIndexedVariable — Type

DoubleIndexedVariable <: CNumber
DoubleIndexedVariable(name::Symbol,ind1::Index,ind2::Index;identical::Bool)

A double-indexed symbolic variable. The variable can (once equations are calculated) be easily exchanged for numerical values.

Fields:

name: A Symbol, defining the name of the variable

ind1: The first Index of the variable

ind2: The second Index of the variable

identical: A Bool, defining if the variable can have non-zero main-diagonal terms, e.g: Γᵢᵢ ≠ 0 would be specified with true.

SecondQuantizedAlgebra.SingleSum — Type

SingleSum <: QTerm

Defines a symbolic summation over a term, using one Index entity.

Fields:

term: A multiplication of q-number terms. When the multiplication contains any IndexedOperator with the same index as the summation-index, a symbolic sum will be created.

sum_index: The index, for which the summation will go over.

nonequalindices: (optional) A vector of indices, for which the summation-index can not be equal with.

SecondQuantizedAlgebra.DoubleSum — Type

DoubleSum <: QTerm

Defines a symbolic summation over another SingleSum, using one Index entity. This corresponds to a double-summation over a multiplication of terms.

Fields:

innerSum: A SingleSum entity.

sum_index: The index, for which the (outer) summation will go over.

NEI: (optional) A vector of indices, for which the (outer) summation-index can not be equal with.

SecondQuantizedAlgebra.change_index — Function

change_index(term,from::Index,to::Index)

Exchanges all occurring indices inside the given term, that are equal to the from to the to index.

Examples

change_index(σⱼ²¹,j,i) = σᵢ²¹

change_index(σⱼ²¹ * σᵢ¹²,j,i) = σᵢ²²

SecondQuantizedAlgebra.insert_index — Function

insert_index(term,ind::Index,value::Int)

Function, that inserts an integer value for a index in a specified term. This function creates Numbered- Variables/Operators/Sums upon calls.

Examples

insert_index(σⱼ²¹,j,1) = σ₁²¹

SecondQuantizedAlgebra.SpecialIndexedTerm — Type

SpecialIndexedTerm <: QTerm

A multiplication of IndexedOperator entities, with special constraint on the index-values. For example σᵢ²² * σⱼ²² with the constraint i ≠ j

Fields:

term: A multiplication of q-number terms.

indexMapping: A Vector of Index tuples, specifying the constraints for the term. Each Tuple is considered to one constraint. e.g: (i,j) -> i ≠ j

SecondQuantizedAlgebra.reorder — Function

reorder(param,indexMapping)

Reorders a given term (param) regarding to a given indexMapping, which specifies, which Index entities can not be equal inside the given term. reorder() creates a SpecialIndexedTerm as a result.

Examples

reorder(σⱼ²¹ * σᵢ²¹,[(i,j)]) = σᵢ²¹ * σⱼ²¹

reorder(σⱼ²¹ * σᵢ²¹ * σⱼ¹²,[(i,j)]) = σᵢ²¹ * σⱼ²²

SecondQuantizedAlgebra.IndexedAverageSum — Type

IndexedAverageSum <: CNumber

Defines a symbolic summation over an average, or a multiplication of several averages, using one Index entity.

Fields:

term: A multiplication of average terms.

sum_index: The index, for which the summation will go over.

nonequalindices: (optional) A vector of indices, for which the summation-index can not be equal with.

SecondQuantizedAlgebra.IndexedAverageDoubleSum — Type

IndexedAverageDoubleSum <: CNumber

Defines a symbolic summation over an IndexedAverageSum, using a Index entity. This schematically represent a double-sum over a multiplication of averages.

Fields:

innerSum: An IndexedAverageSum entity.

sum_index: The index, for which the (outer) summation will go over.

nonequalindices: (optional) A vector of indices, for which the (outer) summation-index can not be equal with.

SecondQuantizedAlgebra.NumberedOperator — Type

NumberedOperator <: QSym

Defines an operator, associated with a Number. Commutator-relations are calculated using these numbers, as a sort of a specific index-value.

Fields:

op: An Operator, either a Transition, a Destroy or a Create can be defined.

numb: An Integer Number.